de-nat-3-the-spanish-decaffeination-final-project-main/requirements-lambda.txt
psycopg2-binary==2.9.10
jmespath==1.0.1
pandas==2.2.3
